Fix restart command

+++ mybot.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
-kill $(cat ./pid.txt)
+if [ "$1" != "--nokill" ]; then
+    kill $(cat ./pid.txt)
+fi
+
 bash -c 'nohup node main.js &>./mybot.log & echo $! >./pid.txt && jobs -p %1'
